+++ src/random.h
@@ -9,7 +9,7 @@ int_fast32_t ranNum(int_fast32_t, int_fast32_t);
 
 void newSeed();
 
-void randomPw(char **, bool, bool, bool, bool, size_t);
+void randomPw(char **, bool, bool, bool, bool, uint_fast32_t);
 
 void outputPw(char *, char *);
