+++ src/stringfunctions.h
@@ -2,6 +2,7 @@
 #define STRINGFUNCTIONS_H
 
 #include <stdlib.h>
+#include <stdint.h>
 
 void outputStringLen(char *);
 size_t stringLen(char *);
@@ -15,10 +16,10 @@ void stringLower(char *, char **);
 void outputStringReverse(char *);
 void stringReverse(char *, char **);
 
-void outputStringTimes(char *, size_t);
-void stringTimes(char *, size_t, char **);
+void outputStringTimes(char *, uint_fast32_t);
+void stringTimes(char *, uint_fast32_t, char **);
 
-void outputStringSplitNum(char *, size_t);
+void outputStringSplitNum(char *, uint_fast32_t);
 
 void outputStringSplitChar(char *, char *);
 
@@ -29,6 +30,6 @@ void outputStringSplitAfterChar(char *, char *);
 void outputStringRemove(char *, char *);
 void stringRemove(char *, char *, char **);
 
-void outputLineAt(char *, size_t);
+void outputLineAt(char *, uint_fast32_t);
 
 #endif
